Made a new package json file for the unit tests
When I tried to merge the PR #223 , I got a merge conflict in the package.json file.
Since I counln't risk screwing up the package.json file, I decided to create a new package.json file for the unit tests.
It's called test_package.json and is located in the root directory of the project.
Since unit test cases are not as important as actually running the project, let's just put these here for now to avoid unwanted trouble.
We can come back to these later.